Another local CSS update.
I added a "home" glyphicon because it's not instantly clear for the user
that they can go to the home page when clicking the logo in the
navigation. I also rearranged "book.html" page for cleaner layout. Added
the "banderoll" at the bottom plus made the button and book image
respond to media queries.