Another local CSS update.
Added media queries and edited the "home" glyphicon so that it's
responsive to even the smallest screens. The content on the blog at the
home page is now edited to that it takes up optimal amount of space, as
well as overall touch up. Next thing is going to be the hardest. The
play inspiration page.